Include most of the desired fields in the JSON
There are other fields in the XML that might be interesting, but this grabs the
majority. Also, some fields are numerical types, and ought to be converted to
the actual value.  The abstracts have to come from a different file, but they're
keyed on the application ID, which is included in the main file.